Интерфейсы вынесены в отдельный файл types.ts
